+++ src/app/api/chat/google/route.ts
@@ -13,19 +13,7 @@ import { POST as UniverseRoute } from '../[provider]/route';
 // so if you want to use with proxy, you need comment the code below
 export const runtime = 'edge';
 
-export const preferredRegion = [
-  'bom1',
-  'cle1',
-  'cpt1',
-  'gru1',
-  'hnd1',
-  'iad1',
-  'icn1',
-  'kix1',
-  'pdx1',
-  'sfo1',
-  'sin1',
-  'syd1',
-];
+// due to gemini-1.5-pro only can be used in us, so we need to set the preferred region only in US
+export const preferredRegion = ['cle1', 'iad1', 'pdx1', 'sfo1'];
 
 export const POST = async (req: Request) => UniverseRoute(req, { params: { provider: 'google' } });